Added defaults to watch for the tutorial screen because it was setting the first launch to true before we go to the main screen causing the welcome texts and images not to be shown. Therefore we added a new value to save in the user defaults for the tutorial itself.
Removed a lot of commented code and useless logging to keep the logs clean.

Changed the predicate for the download of the themes so that it checks if it contains the path when selecting the random images to add to the array once we have less than 10 left to download so that we always have 10 in the swipe view.